add admin/partners pages
