+++ apps/web/app/(ee)/admin.dub.co/(dashboard)/layout-nav-client.tsx
@@ -36,10 +36,6 @@ const tabs = [
     href: "/revenue",
     label: "Revenue",
   },
-  {
-    href: "/fraud-alerts",
-    label: "Fraud Alerts",
-  },
 ];
 
 export function AdminNav() {
