+++ apps/web/app/(ee)/admin.dub.co/(dashboard)/partners/layout.tsx
@@ -1,0 +1,14 @@
+import { PartnersNavTabs } from "./partners-nav-tabs";
+
+export default function AdminPartnersLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <div className="mx-auto flex w-full max-w-screen-xl flex-col gap-6 p-6">
+      <PartnersNavTabs />
+      {children}
+    </div>
+  );
+}
