+++ apps/web/app/(ee)/admin.dub.co/(dashboard)/partners/partners-nav-tabs.tsx
@@ -1,0 +1,35 @@
+"use client";
+
+import { PageNavTabs } from "@/ui/layout/page-nav-tabs";
+import { AlertCircleFill } from "@/ui/shared/icons";
+import { GridIcon, UserCheck } from "@dub/ui";
+import { useMemo } from "react";
+
+export function PartnersNavTabs() {
+  const tabs = useMemo(
+    () => [
+      {
+        id: "trusted",
+        label: "Trusted Partners",
+        icon: UserCheck,
+      },
+      {
+        id: "network",
+        label: "Network Applications",
+        icon: GridIcon,
+      },
+      {
+        id: "fraud",
+        label: "Fraud Alerts",
+        icon: AlertCircleFill,
+      },
+    ],
+    [],
+  );
+
+  return (
+    <div className="rounded-xl border border-neutral-200 bg-neutral-100">
+      <PageNavTabs basePath="/partners" tabs={tabs} />
+    </div>
+  );
+}
